Initial import from OpenStack Horizon
Since we are using OpenStack Keystone as our authentication backend, we
can use a lot of the upstream gui too.  Sadly, we have to embed a lot of
the code into our project.  However, there is working happening upstream
to allow project to extend the Keystone web logic.  Once that happens,
we can drop a lot of embedded stuff.

Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>